PWMAudioStreamPico

+++ src/AudioPWM/PWMAudioBase.h
@@ -221,10 +221,10 @@ class PWMAudioStreamBase : public AudioPrint, public AudioBaseInfoDependent {
         // blocking write for an array: we expect a singed value and convert it into a unsigned 
         virtual size_t write(const uint8_t *wrt_buffer, size_t size){
             size_t available = min((size_t)availableForWrite(),size);
-            LOGD("write: %zu bytes -> %zu", size, available);
+            LOGD("write: %u bytes -> %u", (unsigned int)size, (unsigned int)available);
             size_t result = buffer->writeArray(wrt_buffer, available);
             if (result!=available){
-                LOGW("Could not write all data: %d -> %d", size, result);
+                LOGW("Could not write all data: %u -> %d", (unsigned int) size, result);
             }
             // activate the timer now - if not already done
             startTimer();
